// frontend/src/app/shared/generated-api/model/backupJob.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BackupJob { 
    id: string;
    createdByUserId: string;
    state: BackupJobStateEnum;
    createdAt: string;
    startedAt?: string;
    completedAt?: string;
    outputIdentifier?: string;
    byteSize?: number;
    checksum?: string;
    errorMessage?: string;
}
export enum BackupJobStateEnum {
    Queued = 'QUEUED',
    Running = 'RUNNING',
    Succeeded = 'SUCCEEDED',
    Failed = 'FAILED',
    Canceled = 'CANCELED'
};